Fixed loading of edit form loading from pin and evac center pins
Updated populate edit form to search for the correct iframe to load the edit form. Some pins dont load the edit form, may be due to issues in trying to access the data. Also fixed centering of form

Updated index.js to move the logic for evac centers into the updateRiskButton function as clearMarkers removes all the evac centers everytime the risk type is changed